+++ ktrain/CHANGELOG.md
@@ -12,10 +12,13 @@ Most recent releases are shown at the top. Each release shows:
 ### New:
 - **cycle\_momentum** argument for both```autofit``` and ```fit_onecycle``` method that will cycle momentum between 0.95 and 0.85 as described in [this paper](https://arxiv.org/abs/1803.09820)
 - ```Learner.plot``` method that will plot training-validation loss, LR schedule, or momentum schedule
+- added ```set_weight_decay``` and ```get_weight_decay``` methods to get/set "global" weight decay in Keras
+- 
 
 ### Changed:
 - ```vision.data.preview_data_aug``` now displays images in rows by default
 - added multigpu flag to ```core.get_learner``` with comment that it is only supported by```vision.model.image_classifier```
+- added ```he_normal``` initialization to FastText model
 
 ### Fixed: